requirements.txt
alembic==1.9.2
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.0
distlib==0.3.6
fastapi==0.89.1
filelock==3.9.0
flower==1.2.0
gitdb==4.0.10
GitPython==3.1.30
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
humanize==4.5.0
identify==2.5.16
idna==3.4
iniconfig==2.0.0
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
nodeenv==1.7.0
orjson==3.8.5
packaging==23.0
platformdirs==2.6.2
pluggy==1.0.0
pre-commit==3.0.1
prometheus-client==0.16.0
prompt-toolkit==3.0.36
pycparser==2.21
pydantic==1.10.4
pytest==7.2.1
pytest-asyncio==0.20.3
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
rabbitmq==0.2.0
redis==4.4.2
rfc3986==1.5.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.0
starlette==0.22.0
tornado==6.2
types-pyOpenSSL==23.0.0.2
types-redis==4.4.0.4
typing_extensions==4.4.0
uvicorn==0.20.0
vine==5.0.0
virtualenv==20.17.1
wcwidth==0.2.6
whatthepatch==1.0.4
